refactor(services): seperate mutations and queries into seperate files
