+++ apps/frontend/src/services/admin/user/AdminUserMutation.ts
@@ -1,0 +1,54 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query"
+import { QueryKeys } from "@/services"
+import AdminUserService from "./AdminUserService"
+
+/**
+ * Provides a mutation hook for creating a new user and invalidates the user query cache.
+ *
+ * @returns A mutation hook that creates a user.
+ */
+export const useCreateUser = () => {
+  const queryClient = useQueryClient()
+  return useMutation({
+    mutationFn: AdminUserService.createUser,
+    // TODO: when get by id is implemented, do not invalidate get by id queries
+    onSuccess: () =>
+      queryClient.invalidateQueries({
+        queryKey: [QueryKeys.USER_QUERY_KEY],
+      }),
+  })
+}
+
+/**
+ * Provides a mutation hook for updating a user and invalidates the user query cache.
+ *
+ * @returns A mutation hook that updates a user.
+ */
+export const useUpdateUser = () => {
+  const queryClient = useQueryClient()
+  return useMutation({
+    mutationFn: AdminUserService.updateUser,
+    // TODO: When get by id is implemented, only invalidate the one id for get by id or update the specific query with the updated data
+    onSuccess: () =>
+      queryClient.invalidateQueries({
+        queryKey: [QueryKeys.USER_QUERY_KEY],
+      }),
+  })
+}
+
+/**
+ * Provides a mutation hook for deleting a user and invalidates the user query cache.
+ *
+ * @returns A mutation hook for deleting a user.
+ */
+export const useDeleteUser = () => {
+  const queryClient = useQueryClient()
+  return useMutation({
+    mutationFn: AdminUserService.deleteUser,
+    // TODO: When get by id is implemented, only invalidate the one id for get by id
+    onSuccess: () =>
+      queryClient.invalidateQueries({
+        queryKey: [QueryKeys.USER_QUERY_KEY],
+      }),
+  })
+}
